FIX: Requisito 1 - Fixando senha secreta
