+++ back-end/package.json
@@ -19,7 +19,7 @@
     "@types/jest": "25.1.4",
     "body-parser": "1.19.0",
     "cors": "2.8.5",
-    "dotenv": "8.2.0",
+    "dotenv": "^8.2.0",
     "eslint-config-trybe-backend": "1.0.3",
     "express": "4.17.1",
     "express-rescue": "1.1.26",
