+++ back-end/src/auth/validateToken.js
@@ -2,21 +2,23 @@ const jwt = require('jsonwebtoken');
 const httpResponse = require('../utils/httpResponses');
 const httpStatusCode = require('../utils/httpStatusCode');
 
-const { SECRET } = process.env;
+const SECRET = "my&7Ip$xk6PIsDL";
+//const { SECRET } = process.env;
 
-module.exports = async (req, res, next) => {
-    const token = req.headers.authorization;
+const validateToken = async (req, res, next) => {
+  const token = req.headers.authorization;
 
-   if (!token) return httpResponse.UNAUTHORIZED; 
+  if (!token) return httpResponse.UNAUTHORIZED; 
 
-    jwt.verify(token, SECRET, (err, decoded) => {
-      if (err) {
-        return res
-          .status(httpStatusCode.UNAUTHORIZED)
-          .json({ message: err.message });
+  jwt.verify(token, SECRET, (err, decoded) => {
+    if (err) {
+      return res
+        .status(httpStatusCode.UNAUTHORIZED)
+        .json({ message: err.message });
       } 
       req.user = decoded;
-    });
-
-    next();
+  });
+  next();
 };
+
+module.exports = validateToken;
