+++ front-end/package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "9.5.0",
     "@testing-library/user-event": "7.2.1",
     "axios": "0.19.2",
+    "dotenv": "^8.2.0",
     "eslint-config-trybe-frontend": "1.2.0",
     "jquery": "1.9.1",
     "react": "16.13.1",
