+++ package.json
@@ -29,7 +29,7 @@
   "devDependencies": {
     "cypress": "5.0.0",
     "cypress-multi-reporters": "1.2.3",
-    "dotenv": "8.2.0",
+    "dotenv": "^8.2.0",
     "faker": "4.1.0",
     "mocha": "7.0.1",
     "mochawesome": "4.1.0",
